requirements.txt
aiohttp==3.8.1
aiohttp-pydantic==1.12.1
aiohttp-sqlalchemy==0.34.0
alembic==1.7.6
SQLAlchemy==1.4.31
swagger-ui-bundle==0.0.9
mysqlclient==2.1.0
aiohttp-jwt==0.6.1
passlib==1.7.4
